+++ src/siteConfig.json
@@ -16,7 +16,7 @@
     "Integrate with Fiji, Ilastik, ImJoy",
     "Try model instantly with BioEngine",
     "Contribute your models via Github",
-    "Link models to BioEngine Apps"
+    "Link models to datasets and applications"
   ],
   "item_lists": [
     {
