+++ src/utils.js
@@ -4,6 +4,23 @@ export function randId() {
     .substr(2, 10);
 }
 
+export function validateBioEngineApp(name, api) {
+  if (!api.runOneModel && !api.runManyModels) {
+    console.error(
+      `${name}" has neither "runOneModel" nor "runManyModels":`,
+      api
+    );
+    alert(
+      `"${name}" is not a valid BioEngine App, it should define "runOneModel" and/or "runManyModels".`
+    );
+    return false;
+  }
+  if (!api.testModel) {
+    console.warn(`Please define a testModel function for "${name}".`);
+  }
+  return true;
+}
+
 export function debounce(func, wait, immediate) {
   var timeout;
   return function() {
